requirements.txt
asgiref==3.11.0
Django==5.0.14
djangorestframework==3.16.1
mysqlclient==2.2.7
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
